+++ src/index.mjs
@@ -1,0 +1,6 @@
+import fetchMock from './index.js'
+
+export default fetchMock
+export const enableFetchMocks = fetchMock.enableFetchMocks
+export const disableFetchMocks = fetchMock.disableFetchMocks
+export const createFetchMock = fetchMock.createFetchMock