+++ src/setup.js
@@ -1,0 +1,2 @@
+// convenience setup file: "setupFiles": ["jest-fetch-mock/setup"]
+require('./index.js').enableMocks()