+++ types/factory.d.ts
@@ -1,0 +1,1 @@
+export { createFetchMock, createFetchMock as default, FetchMock, MockFunctionFactory } from "./index";